requirements.txt
Django>=3.0.2,<3.1.0
djangorestframework>=3.11.0,<3.12.0

flake8>=3.7.9,<3.8.0

Pillow>=7.0.0,<7.1.0
phonenumbers>=8.12.0,<8.13.0
django-phonenumber-field>=4.0.0,<4.1.0
